Updated git ignore and working on heat
