--- pysweep/tests/inputs/example.yaml
@@ -1,25 +1,0 @@
----
-  swept: True
-
-  filename: "example.hdf5"
-
-  verbose: True
-
-  blocksize: 8
-
-  share: 0.25
-
-  dtype: float64
-
-  globals:
-    - 0
-    - 10
-    - 0.1
-    - 0.1
-    - 0.1
-
-  intermediate_steps: 1
-  operating_points: 1
-
-  cpu: /home/anthony-walker/nrg-swept-project/pysweep-git/pysweep/equations/example.py
-  gpu: ./pysweep-git/pysweep/equations/example.cu
